Added node failure test and in that processes fixed bug with failure node flow

+++ serac/src/main/java/com/vineeth/serac/manager/pipeline/processors/EvaluateFailureNodesProcessor.java
@@ -43,7 +43,7 @@ public class EvaluateFailureNodesProcessor implements IProcessor {
             Node node = nodeStore.getNodeById(nodeId);
             if(node.isHealthy()) {
                 int suspectCount = suspectStore.getSuspectCountForNodeId(nodeId);
-                if(suspectCount / totalNodes > quorumPercentage) {
+                if(((float) suspectCount / totalNodes) > quorumPercentage) {
                     logger.info("Node {} marked FAILED by Node {}", node.getId(), nodeStore.getCurrentNode().getId());
                     node.setHealthy(false);
                 }
